feat(Player): add collision detection and explosion
The player can now collide with objects and an explosion particle is
emitted to show the player's death.

There are a number of tags on objects that prevent the player's death.

The `IgnoreCollision` tag completely omits the collision check and is used
by the game boundaries to prevent destruction. Also the `SafeCollision`
tag is used by objects that the player can land on so will prevent
destruction.