feat(LevelManager): add level map generation
The new `LevelManager` code is a singleton that holds the state of the
level and is responsible for creating the level map.

The level map is a procedurally generated collection of buildings at
different heights with an additional building top that is safe to land
on.

An additional light source has been added to provide directional
shadows as the main light source is directly above the game world.

The `LevelManager` has also been made to reset the player when they
die.