refactor(Player): consolidate rotate animation into one script
There is no need to have two scripts for the player body rotation as
the two scripts are doing the same thing but just have a different
rotational factor which is set as a public variable anyway.

The `PlayerWingAnimation` script has been removed and now both the
player wing and central area use the `PlayerCentralAnimation` script.