feat(Player): prevent player being obscured by buildings
A new `CameraRay` object has been created that extends from the
camera to the player and is used to determine if there is any building
obscuring the player from the camera view.

If there is, then the building block that is in the way is temporarily
unrendered until it is no longer blocking the view.

A new custom shader has been added to address the z-buffer issues when
turning off object renders causes the particles to render in front of
buildings that have had their render mode tampered with.

The new shader does a double render so it fixes this issue.