feat(Peep): add peeps on top of buildings to collect or kill
The `Peep` is the aim of each level, the player must either collect
or kill the peeps in order to progress to the next level. Collecting a
peep requires the player to land successfully on top of the building
the peep is on.

Some additional code has been fixed and tweaked to work correctly with
the new `Peep` mechanic.