feat(GameManager): add game states and increasing levels
The new `GameManager` handles the global game state for player lives,
score, peeps per level, etc.

Also, new game state scenes have been added for a main menu, next level
and game over screen that all utilise the global `GameManager` for
sharing state between scenes.

The new state scenes are very basic at the moment and will be improved
in the future.

The level now increases 1 peep per time each time all peeps are
collected on a level, this is a basic mechanic at the moment but will
be improved in the future.

The player also now has a score which is kept through each level via
the `GameManager`.