feat(GameManager): add screen fade between scene loading
The `GameManager` now handles a screen fade between loading each scene
which makes the transition between scenes look neater.

Simply calling the `LoadScene(<name>)` method on the `GameManager` is
all that is required to enable the fading between loading.

A new black pixel texture has been added that is used for the GUI fade.